requirements.txt
markdown
jinja2
livereload
python-frontmatter
pygments
pymdown-extensions
Pillow
